\name{nNodes-methods}
\docType{methods}
\alias{nNodes-methods}
\alias{nTips-methods}
\alias{print-methods}
\alias{nNodes,phylo4-method}
\alias{nTips,phylo4-method}
\title{Methods for S4 phylog classes}
\description{
   Generic methods for phylogenetic trees represented as S4 classes
}
\section{Methods}{
\describe{

  \item{nNodes}{\code{signature(object="phylo4")}: extract number of
    nodes}
  \item{nTips}{\code{signature(object="phylo4")}: extract number of tips}
}
}
\keyword{methods}
